requirements/requirements.txt
Click==7.0
cycler==0.10.0
fastapi==0.70.0
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
pika==1.2.0
Pillow==6.2.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-multipart==0.0.5
PyYAML==6.0.0
redis==3.5.3
sentry-sdk==1.3.1
six==1.13.0
uvicorn==0.15.0
Werkzeug==0.16.0
